fix: refactoring(add cntrl)
