+++ helpers/cntrlWrapper.js
@@ -1,0 +1,13 @@
+const cntrlWrapper = (cntrl) => {
+  const func = async (req, res, next) => {
+    try {
+      await cntrl(req, res, next);
+    } catch (err) {
+      next(err);
+    }
+  };
+
+  return func;
+};
+
+module.exports = cntrlWrapper;
