+++ helpers/index.js
@@ -1,5 +1,7 @@
 const HttpError = module.require("./HttpError");
+const cntrlWrapper = module.require("./cntrlWrapper");
 
 module.exports = {
   HttpError,
+  cntrlWrapper,
 };
