+++ models/contacts.json
@@ -1,7 +1,7 @@
 [
   {
     "id": "vza2RIzNGIwutCVCs4mCL",
-    "name": "Wylie Wonka",
+    "name": "Wylie",
     "email": "est@utquamvel.net",
     "phone": "(692) 802-2949"
   },
@@ -52,5 +52,23 @@
     "email": "gaailghg@gmail.com",
     "phone": "+380996456442",
     "id": "86a4066b-4ed6-42c9-b413-02d6e820804c"
+  },
+  {
+    "name": "Puyol",
+    "email": "dfdwawr@gmail.com",
+    "phone": "+38099234242",
+    "id": "a427695c-4968-48c4-b688-3af1369228cb"
+  },
+  {
+    "name": "Ronaldo",
+    "email": "dadqwrf@mail.com",
+    "phone": "+38099234242",
+    "id": "9a16defb-a065-4241-9c79-d5428062ec74"
+  },
+  {
+    "name": "Rivaldo",
+    "email": "padrtgfdsf@mail.com",
+    "phone": "+38099234242",
+    "id": "c85bbacf-9027-44e3-be2e-c4a5779881a6"
   }
 ]
